backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
opencv-python-headless
onnxruntime
httpx
openai
tenacity
pillow
numpy
scikit-image
python-multipart
aiofiles
